ChatGPT4QAZ/requirements.txt
aiogram==2.25.1
openai==0.27.8
asyncpg==0.28.0
APScheduler==3.6.3
pytz==2023.3
python-dateutil==2.8.2